+++ plugin_tooling/src-util/melnorme/utilbox/collections/HashMap2.java
@@ -20,7 +20,7 @@ import melnorme.utilbox.core.CoreUtil;
  * Extension to {@link java.util.HashMap}, with some helper methods,
  * and implementing the read-only interface {@link Collection2}.
  */
-public class HashMap2<K, V> extends HashMap<K, V> implements Collection2<Map.Entry<K, V>> {
+public class HashMap2<K, V> extends HashMap<K, V> implements MapAccess<K, V> {
 	
 	private static final long serialVersionUID = -7612795787860334443L;
 	
@@ -30,6 +30,11 @@ public class HashMap2<K, V> extends HashMap<K, V> implements Collection2<Map.Ent
 	
 	public HashMap2(Map<? extends K, ? extends V> map) {
 		super(map);
+	}
+	
+	@Override
+	public HashMap2<K, V> copyToHashMap() {
+		return copy();
 	}
 	
 	public HashMap2<K, V> copy() {